Update code from GitLab

+++ structs/educations.struct.go
@@ -7,7 +7,7 @@ type Education struct {
 	UserProfileId       int      `json:"user_profile_id"`
 	Description         string   `json:"description"`
 	DateOfCertification JSONDate `json:"date_of_certification"`
-	Price               int      `json:"price"`
+	Price               float32  `json:"price"`
 	DateOfStart         JSONDate `json:"date_of_start"`
 	DateOfEnd           JSONDate `json:"date_of_end"`
 	AcademicTitle       string   `json:"academic_title"`
